refactor: move loop-body and declaration finders to core ast-node finders
