+++ packages/eslint/src/rules/prevent-construct-id-collision.ts
@@ -1,5 +1,7 @@
 import { AST_NODE_TYPES, ESLintUtils, TSESTree } from "@typescript-eslint/utils";
 
+import { isDeclaredInEnclosingBlocks } from "../core/ast-node/finder/declared-in-enclosing-blocks";
+import { findEnclosingLoopBody } from "../core/ast-node/finder/enclosing-loop-body";
 import { isConstructType } from "../core/cdk-construct/type-checker/is-construct";
 import { findConstructorPropertyNames } from "../core/ts-type/finder/constructor-property-name";
 import { createRule } from "../shared/create-rule";
@@ -81,44 +83,6 @@ const validateConstructIdInLoop = (
 };
 
 /**
- * Find the body of the loop that encloses a node.
- * Detects for, for...in, for...of, while, do...while statements,
- * and callbacks of iteration methods (forEach, map, etc.)
- * Returns null when the node is not inside a loop.
- */
-const findEnclosingLoopBody = (node: TSESTree.Node): TSESTree.Node | null => {
-  const parent = node.parent;
-  if (!parent) return null;
-
-  // NOTE: Detect loop statements
-  if (
-    parent.type === AST_NODE_TYPES.ForStatement ||
-    parent.type === AST_NODE_TYPES.ForInStatement ||
-    parent.type === AST_NODE_TYPES.ForOfStatement ||
-    parent.type === AST_NODE_TYPES.WhileStatement ||
-    parent.type === AST_NODE_TYPES.DoWhileStatement
-  ) {
-    return parent.body;
-  }
-
-  // NOTE: Detect iteration method callbacks (ArrowFunction/FunctionExpression)
-  if (
-    (parent.type === AST_NODE_TYPES.ArrowFunctionExpression ||
-      parent.type === AST_NODE_TYPES.FunctionExpression) &&
-    isIterationMethodCallback(parent)
-  ) {
-    return parent.body;
-  }
-
-  // NOTE: Stop at non-constructor method definitions
-  if (parent.type === AST_NODE_TYPES.MethodDefinition && parent.kind !== "constructor") {
-    return null;
-  }
-
-  return findEnclosingLoopBody(parent);
-};
-
-/**
  * Check whether the scope argument is an identifier declared inside the loop body.
  * Only a variable declaration is treated as a per-iteration scope: any other form
  * (`this`, an outer variable, a member expression, ...) may be shared across iterations.
@@ -129,66 +93,4 @@ const isScopeDeclaredInLoopBody = (
 ): boolean => {
   if (scopeArg.type !== AST_NODE_TYPES.Identifier) return false;
   return isDeclaredInEnclosingBlocks(scopeArg.name, scopeArg, loopBody);
-};
-
-/**
- * Check whether a block enclosing the node, up to and including the loop body,
- * declares a variable with the given name
- */
-const isDeclaredInEnclosingBlocks = (
-  name: string,
-  node: TSESTree.Node,
-  loopBody: TSESTree.Node,
-): boolean => {
-  if (declaresVariableName(node, name)) return true;
-  if (node === loopBody || !node.parent) return false;
-  return isDeclaredInEnclosingBlocks(name, node.parent, loopBody);
-};
-
-/**
- * Check whether a block statement declares a variable with the given name
- */
-const declaresVariableName = (node: TSESTree.Node, name: string): boolean => {
-  if (node.type !== AST_NODE_TYPES.BlockStatement) return false;
-
-  return node.body.some(
-    (statement) =>
-      statement.type === AST_NODE_TYPES.VariableDeclaration &&
-      statement.declarations.some(
-        (declarator) =>
-          declarator.id.type === AST_NODE_TYPES.Identifier && declarator.id.name === name,
-      ),
-  );
-};
-
-const ITERATION_METHODS = new Set([
-  "forEach",
-  "map",
-  "flatMap",
-  "filter",
-  "reduce",
-  "reduceRight",
-  "every",
-  "some",
-  "find",
-  "findIndex",
-  "findLast",
-  "findLastIndex",
-]);
-
-/**
- * Check whether an arrow function or function expression is a callback of an iteration method
- */
-const isIterationMethodCallback = (
-  node: TSESTree.ArrowFunctionExpression | TSESTree.FunctionExpression,
-): boolean => {
-  const parent = node.parent;
-  if (parent?.type !== AST_NODE_TYPES.CallExpression) return false;
-
-  const callee = parent.callee;
-  if (callee.type !== AST_NODE_TYPES.MemberExpression) return false;
-
-  if (callee.property.type !== AST_NODE_TYPES.Identifier) return false;
-
-  return ITERATION_METHODS.has(callee.property.name);
 };
